+++ benchmarks/benchmark-recint_mul.C
@@ -19,8 +19,7 @@ using namespace RecInt;
 
 int main(int argc, char ** argv)
 {
-    size_t nbloops = (argc>1?atoi(argv[1]):LOOPS);
-    
+    size_t nbloops = static_cast<size_t>((argc > 1)? atoi(argv[1]) : LOOPS);
 
     rmint<STD_RECINT_SIZE> a[ALEA_MAX];
     ruint<STD_RECINT_SIZE> module;
@@ -41,8 +40,7 @@ int main(int argc, char ** argv)
     for (unsigned int l = 0; l < nbloops; l++) {
         mul(a[l & ALEA_MASK], a[l & ALEA_MASK], a[(l+1) & ALEA_MASK]);
     }
-    tim.stop();
-    
+    tim.stop(); 
     
 	// -----------
 	// Standard output for benchmark - Alexis Breust 2014/12/11
